Hide LLVM from header files
