+++ src/AggregateFunctions/IAggregateFunction.cpp
@@ -132,8 +132,6 @@ void IAggregateFunction::predictValues(
     throw Exception(ErrorCodes::NOT_IMPLEMENTED, "Method predictValues is not supported for {}", getName());
 }
 
-#if USE_EMBEDDED_COMPILER
-
 void IAggregateFunction::compileCreate(llvm::IRBuilderBase & /*builder*/, llvm::Value * /*aggregate_data_ptr*/) const
 {
     throw Exception(ErrorCodes::NOT_IMPLEMENTED, "{} is not JIT-compilable", getName());
@@ -156,5 +154,4 @@ llvm::Value * IAggregateFunction::compileGetResult(llvm::IRBuilderBase & /*build
     throw Exception(ErrorCodes::NOT_IMPLEMENTED, "{} is not JIT-compilable", getName());
 }
 
-#endif
 }
